requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-DebugToolbar==0.14.1
Flask-Login==0.4.1
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==1.1.0
Jinja2==2.11.3
loguru==0.6.0
MarkupSafe==1.1.1
mysqlclient==2.2.1
packaging==23.2
PyMySQL==0.9.3
pyodbc==5.0.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==0.15.6
